Melhoria dos nomes das váriaveis; Melhoria dos códigos do Decorator da classe AdicionalLanche; Implementação das fábricas dos ingredientes

+++ Lanchonete/src/Ingredientes/Hamburger.java
@@ -1,18 +1,14 @@
 package Ingredientes;
 
-public class Hamburger extends Recheio{
+public class Hamburger extends Recheio {
 
-        
-    
-    public Hamburger(){
-        descricao = ",Hamburger";
-    }
-    
-    
-    @Override
-    public float calcularPreco(){
-        return (float) 12.50;
-    }
-    
-    
+	public Hamburger() {
+		descricao = ",Hamburger";
+	}
+
+	@Override
+	public double calcularPreco() {
+		return (float) 12.50;
+	}
+
 }
