+++ Lanchonete/src/Model/beam/Produto.java
@@ -2,10 +2,10 @@ package Model.beam;
 
 public abstract class Produto {
 
-    float preco;
+    double preco;
     String tipo;
     String descricao;
     
     public abstract String getDescricao();
-    public abstract float getPreco();
+    public abstract double getPreco();
 }
